+++ carts/models.py
@@ -1,6 +1,6 @@
+from decimal import Decimal
 from django.db import models
-from rest_framework import serializers
-from rest_framework.serializers import ModelSerializer
+
 
 from users.models import User
 from items.models import Item
@@ -9,37 +9,24 @@ from items.models import Item
 
 
 class Cart(models.Model):
-    items = models.ManyToManyField(Item)
+    items = models.ManyToManyField(to=Item, through='CartItem')
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='carts')
+
+    total_cost = models.DecimalField(max_digits=13, decimal_places=2, default=True)
+
+    @property
+    def total_cost(self):
+        return self.carts.price * Decimal(self.carts.quantity)
 
 
 class CartItem(models.Model):
-    item = models.ForeignKey(Item, on_delete=models.CASCADE, related_name='cartitems')
-    cart = models.ForeignKey(Cart, on_delete=models.CASCADE, related_name='cartitems')
+    item = models.ForeignKey(to=Item, on_delete=models.CASCADE, related_name='cart_items')
+    cart = models.ForeignKey(to=Cart, on_delete=models.CASCADE, related_name='cart_items')
     quantity = models.PositiveIntegerField()
     price = models.DecimalField(max_digits=13, decimal_places=2)
 
+    total_price = models.DecimalField(max_digits=13, decimal_places=2, default=True)
 
-class CartSerializer(ModelSerializer):
-    total_cost = serializers.SerializerMethodField()
-
-    def get_total_cost(self, cart_instance):
-        return cart_instance.a + cart_instance.b
-
-    class Meta:
-        model = Cart
-        fields = ('items', 'user', 'total_cost')
-
-
-class CartItemSerializer(ModelSerializer):
-    total_price = serializers.SerializerMethodField()
-
-    def get_total_price(self, cart_item_instance):
-        return cart_item_instance.a + cart_item_instance.b
-
-    class Meta:
-        model = CartItem
-        fields = ('item', 'cart', 'quantity', 'price', 'item_id', 'total_price')
-        extra_kwargs = {
-            'price': {'read_only': True},
-        }
+    @property
+    def total_price(self):
+        return self.price * Decimal(self.quantity)
