+++ food_box/urls.py
@@ -43,6 +43,5 @@ api_url = [
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/v1/', include(urlpatterns_users)),
     path('api/v1/', include(api_url)),
 ]
